splited MetaDataManager
