+++ src/main/java/com/custom/orm/metadata/implementation/MetaDataManagerImpl.java
@@ -1,0 +1,6 @@
+package com.custom.orm.metadata.implementation;
+
+import com.custom.orm.metadata.MetaDataManager;
+
+public class MetaDataManagerImpl extends MappingMetaDataImpl implements MetaDataManager {
+}
